@@ -1,30 +1,30 @@
-<template>
-  <div id="app">
-    <h1>SuccessDomino</h1>
-    <ChallengeForm />
-  </div>
-</template>
-
-<script>
-// import HelloWorld from './components/HelloWorld.vue'
-import ChallengeForm from './components/ChallengeForm.vue';
-
-
-export default {
-  name: 'App',
-  components: {
-    ChallengeForm
-  }
-}
+<script setup lang="ts">
+import ChallengeForm from '@/components/ChallengeForm.vue'
 </script>
 
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
+<template>
+  <div>
+    <a href="https://vite.dev" target="_blank">
+      <img src="/vite.svg" class="logo" alt="Vite logo" />
+    </a>
+    <a href="https://vuejs.org/" target="_blank">
+      <img src="./assets/vue.svg" class="logo vue" alt="Vue logo" />
+    </a>
+  </div>
+  <ChallengeForm />
+</template>
+
+<style scoped>
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
+}
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+.logo.vue:hover {
+  filter: drop-shadow(0 0 2em #42b883aa);
 }
 </style>
